Use BigInteger for Discord entity IDs

+++ event_bot/models/event.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Integer, Text, ForeignKey
+from sqlalchemy import BigInteger, Column, DateTime, Integer, Text, ForeignKey
 from sqlalchemy.orm import relationship
 
 from . import Base
@@ -6,9 +6,9 @@ from . import Base
 class Event(Base):
     __tablename__ = 'event'
     id = Column(Integer, primary_key=True)
-    event_channel_id = Column(Integer, ForeignKey('event_channel.id', ondelete='CASCADE'))
-    user_id = Column(Integer, ForeignKey('user.id', ondelete='CASCADE'))
-    message_id = Column(Integer)
+    event_channel_id = Column(BigInteger, ForeignKey('event_channel.id', ondelete='CASCADE'))
+    user_id = Column(BigInteger, ForeignKey('user.id', ondelete='CASCADE'))
+    message_id = Column(BigInteger)
     title = Column(Text)
     description = Column(Text)
     start_time = Column(DateTime)
