+++ event_bot/models/event_channel.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, BigInteger, ForeignKey
 from sqlalchemy.orm import relationship
 
 from . import Base
 
 class EventChannel(Base):
     __tablename__ = 'event_channel'
-    id = Column(Integer, primary_key=True)
-    guild_id = Column(Integer, ForeignKey('guild.id', ondelete='CASCADE'))
+    id = Column(BigInteger, primary_key=True)
+    guild_id = Column(BigInteger, ForeignKey('guild.id', ondelete='CASCADE'))
 
     events = relationship("Event", passive_deletes=True)
